Add files via upload
A simple program to play 3 rounds of rock, paper, scissor, with the computer. With a print_slow process to make it seem like it is a human writing the strings. I made a print-based debug system that tells you where in the process there's being an error. If the user gives an invalid answer it restarts the round.